Security kacper

+++ src/main/java/com/codecool/gastro/dto/review/ReviewDto.java
@@ -1,14 +1,13 @@
 package com.codecool.gastro.dto.review;
 
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 public record ReviewDto(
         UUID id,
         String comment,
-        Double grade,
+        int grade,
         LocalDate submissionTime
 ) {
 }
